requirements.txt
langchain
openai
fastapi
uvicorn
python-dotenv
requests